# Enable property functions for "fill-color" and "fill-outline-color"