--- a/js/render/draw_circle.js
+++ b/js/render/draw_circle.js
@@ -25,11 +25,12 @@ function drawCircles(painter, source, layer, coords) {
         var bufferGroups = bucket.bufferGroups.circle;
         if (!bufferGroups) continue;
 
+        var programOptions = bucket.paintAttributes.circle[layer.id];
         var program = painter.useProgram(
             'circle',
-            bucket.paintAttributes.circle[layer.id].defines,
-            bucket.paintAttributes.circle[layer.id].vertexPragmas,
-            bucket.paintAttributes.circle[layer.id].fragmentPragmas
+            programOptions.defines,
+            programOptions.vertexPragmas,
+            programOptions.fragmentPragmas
         );
 
         gl.uniform2f(program.u_extrude_scale,