--- a/package.json
+++ b/package.json
@@ -19,7 +19,7 @@
     "grid-index": "^0.1.0",
     "mapbox-gl-function": "^1.2.1",
     "mapbox-gl-js-supported": "^1.1.0",
-    "mapbox-gl-shaders": "mapbox/mapbox-gl-shaders#556b3ec8ea48c2135d7111ae650dc09bcbd720ef",
+    "mapbox-gl-shaders": "mapbox/mapbox-gl-shaders#e4737bb136d718f9c5fe8d943380f05db6249b57",
     "mapbox-gl-style-spec": "^8.7.0",
     "minifyify": "^7.0.1",
     "pbf": "^1.3.2",