ref https://github.com/mapbox/mapbox-gl-shaders/pull/3
ref https://github.com/mapbox/mapbox-gl-test-suite/pull/99

cc @jakepruitt @jfirebaugh @mourner @tmcw 

<img width="1280" alt="screen shot 2016-05-26 at 2 55 21 pm" src="https://cloud.githubusercontent.com/assets/281306/15593376/728e4936-2360-11e6-8ecc-8fae0a67c99d.png">