@jfirebaugh 2016-05-27
I'm not caught up on gl-js buffer architecture so I don't think I can meaningfully review here.