@lucaswoj 2016-05-27
Main thread perf seems to have taken a hit: 

# Benchmarks

## Buffer

this branch: 854 ms
master: 827 ms

## FPS

this branch: 48fps
master: 59 fps

## Frame Duration

this branch:  19.7 ms per frame. 72% of frames took longer than 16ms.
master: 7.8 ms per frame. 0% of frames took longer than 16ms.