@mourner 2016-05-27
@lucaswoj whoa, 2.5x slower frame is a blocker for merging. I guess we'll have to refactor this.