@lucaswoj 2016-05-27
@mourner the perf regression was due to a bug in `createProgram`. New numbers:

# Benchmarks

## Buffer

this branch: 809 ms
master: 827 ms

## FPS

this branch: 60fps
master: 59 fps

## Frame Duration

this branch:  9.7 ms per frame. 1% of frames took longer than 16ms.
master: 7.8 ms per frame. 0% of frames took longer than 16ms.