@lucaswoj 2016-05-27
@mourner I suspect that 2ms is just measurement error. I'll try to create a cleaner testing environment and see if the discrepancy goes away.