@lucaswoj 2016-05-27
I was wrong -- the difference is definitely significant. 

## this Branch

9.7 ms per frame. 1% of frames took longer than 16ms.
9.8 ms per frame. 3% of frames took longer than 16ms.
10.5 ms per frame. 3% of frames took longer than 16ms.

## master

8.3 ms per frame. 0% of frames took longer than 16ms.
8.2 ms per frame. 2% of frames took longer than 16ms.
7.9 ms per frame. 0% of frames took longer than 16ms.