@lucaswoj 2016-05-27
I got a smidgen of perf back by optimizing `useProgram` further. I suspect that the rest of the perf loss is due to moving [this uniform setup](https://github.com/mapbox/mapbox-gl-js/pull/2629/files#diff-a406b8ffc7f2d9f303442aecebb9a2bdL47) inside a [hotter loop](https://github.com/mapbox/mapbox-gl-js/pull/2629/files#diff-a406b8ffc7f2d9f303442aecebb9a2bdR142).