@mourner 2016-05-27
In context of frame performance, even 2ms is 20% difference. Let's make sure we don't loose track of this, maybe ticket?